tests: initial commit for matrix assembly testing
Changing Makefile in order to run tests on matrix stuff. In this
commit I also refactored a bunch of assembly routines but code there is
still far from nice...

By the looks of it, it seems that local_stiffness is ok. Tests should be
memory leak free but I am not sure.

Signed-off-by: Darko Janekovic <darko.janekovic@fer.hr>