mesh.c: added prealloc nnz array calculation
Array is generated in the loop with signs matrix and discrete gradient
matrix. Same loop is used in favour of speed.

Preallocation is tested with -info command line option, and zero
malloc's have been called.

$ ./bin/eddy -info

Preallocation algorithm
=======================

If some edge is shared, that row in matrix will have coefficients of
both triangles. Not sure what about 3D but idea should generalize. Max
number should be 11, and min number should be 6 in case of nedelec basis
function.

Signed-off-by: Darko Janekovic <darko.janekovic@fer.hr>